requirements.txt
PyNaCl
pillow
pymongo
motor
discord
dnspython
